trekker ut opprettholdelse av vedtak som en egen komponent i vurderingsfanen. Justerer feilmeldinger. Sletter unødvendige states
