+++ src/frontend/Komponenter/Behandling/Brev/Brev.tsx
@@ -7,42 +7,23 @@ import {
     RessursStatus,
     RessursSuksess,
 } from '../../../App/typer/ressurs';
-import { useBehandling } from '../../../App/context/BehandlingContext';
 import { useApp } from '../../../App/context/AppContext';
-import { Alert, Box, Button, HGrid, VStack } from '@navikt/ds-react';
 import { IVurdering, VedtakValg } from '../Vurdering/vurderingValg';
-import PdfVisning from './PdfVisning';
-import { ModalWrapper } from '../../../Felles/Modal/ModalWrapper';
-import SystemetLaster from '../../../Felles/SystemetLaster/SystemetLaster';
-import BrevMottakere from '../Brevmottakere/ef/BrevMottakere';
 import { OmgjørVedtak } from './OmgjørVedtak';
-import { Behandling, Fagsystem } from '../../../App/typer/fagsak';
-import { BrevmottakerContainer as BaksBrevmottakerContainer } from '../Brevmottakere/baks/BrevmottakerContainer';
-import { useFerdigstillBehandling } from './useFerdigstillBehandling';
-
-type Utfall = 'IKKE_SATT' | 'LAG_BREV' | 'OMGJØR_VEDTAK';
+import { Fagsystem } from '../../../App/typer/fagsak';
+import DataViewer from '../../../Felles/DataViewer/DataViewer';
+import { OpprettholdVedtak } from './OpprettholdVedtak';
+import { Alert } from '@navikt/ds-react';
 
 interface Props {
-    behandling: Behandling;
+    behandlingId: string;
+    fagsystem: Fagsystem;
 }
 
-export const Brev: React.FC<Props> = ({ behandling }) => {
-    const behandlingId = behandling.id;
-
-    const [brevRessurs, settBrevRessurs] = useState<Ressurs<string>>(byggTomRessurs());
-
-    const { behandlingErRedigerbar } = useBehandling();
-
+export const Brev: React.FC<Props> = ({ behandlingId, fagsystem }) => {
     const { axiosRequest } = useApp();
-    const { ferdigstill, senderInn } = useFerdigstillBehandling(
-        behandlingId,
-        () => lukkModal(),
-        (feilmelding) => settFeilmelding(feilmelding)
-    );
-    const [visModal, settVisModal] = useState<boolean>(false);
     const [feilmelding, settFeilmelding] = useState('');
-
-    const [utfall, settUtfall] = useState<Utfall>('IKKE_SATT');
+    const [vurdering, settVurdering] = useState<Ressurs<IVurdering | undefined>>(byggTomRessurs());
 
     const hentVurdering = useCallback(
         (behandlingId: string) => {
@@ -51,10 +32,12 @@ export const Brev: React.FC<Props> = ({ behandling }) => {
                 url: `/familie-klage/api/vurdering/${behandlingId}`,
             }).then((response: RessursSuksess<IVurdering | undefined> | RessursFeilet) => {
                 if (response.status === RessursStatus.SUKSESS) {
-                    if (response.data?.vedtak === VedtakValg.OMGJØR_VEDTAK) {
-                        settUtfall('OMGJØR_VEDTAK');
+                    if (!response.data?.vedtak) {
+                        settFeilmelding(
+                            'Det er ikke tatt stilling til om vedtaket skal opprettholdes eller omgjøres. Vennligst naviger til vurderingsfanen for å ta stilling til dette.'
+                        );
                     } else {
-                        settUtfall('LAG_BREV');
+                        settVurdering(response);
                     }
                 } else {
                     settFeilmelding(response.frontendFeilmelding);
@@ -68,81 +51,30 @@ export const Brev: React.FC<Props> = ({ behandling }) => {
         hentVurdering(behandlingId);
     }, [behandlingId, hentVurdering]);
 
-    const hentBrev = useCallback(() => {
-        axiosRequest<string, null>({
-            method: 'GET',
-            url: `/familie-klage/api/brev/${behandlingId}/pdf`,
-        }).then(settBrevRessurs);
-    }, [axiosRequest, behandlingId]);
-
-    const genererBrev = useCallback(() => {
-        axiosRequest<string, null>({
-            method: 'POST',
-            url: `/familie-klage/api/brev/${behandlingId}`,
-        }).then((respons: Ressurs<string>) => {
-            settBrevRessurs(respons);
-        });
-    }, [axiosRequest, behandlingId]);
-
-    useEffect(() => {
-        if (utfall === 'LAG_BREV') {
-            if (behandlingErRedigerbar) {
-                genererBrev();
-            } else {
-                hentBrev();
-            }
-        }
-    }, [behandlingErRedigerbar, genererBrev, hentBrev, utfall]);
-
-    const lukkModal = () => {
-        settVisModal(false);
-        settFeilmelding('');
-    };
-
-    if (utfall === 'LAG_BREV') {
-        return (
-            <Box margin="8">
-                <HGrid gap={'6'} columns={{ xl: 1, '2xl': '1fr 1.2fr' }}>
-                    <VStack gap={'6'}>
-                        {brevRessurs.status === RessursStatus.SUKSESS &&
-                            (behandling.fagsystem !== Fagsystem.EF ? (
-                                <BaksBrevmottakerContainer behandlingId={behandling.id} />
-                            ) : (
-                                <BrevMottakere behandlingId={behandling.id} />
-                            ))}
-                        {behandlingErRedigerbar && brevRessurs.status === RessursStatus.SUKSESS && (
-                            <Button
-                                variant={'primary'}
-                                size={'medium'}
-                                onClick={() => settVisModal(true)}
-                            >
-                                Ferdigstill behandling og send brev
-                            </Button>
-                        )}
-                    </VStack>
-                    <PdfVisning pdfFilInnhold={brevRessurs} />
-                </HGrid>
-                <ModalWrapper
-                    tittel={'Bekreft utsending av brev'}
-                    visModal={visModal}
-                    onClose={() => lukkModal()}
-                    aksjonsknapper={{
-                        hovedKnapp: {
-                            onClick: () => ferdigstill(),
-                            tekst: 'Send brev',
-                            disabled: senderInn,
-                        },
-                        lukkKnapp: { onClick: () => lukkModal(), tekst: 'Avbryt' },
-                    }}
-                    ariaLabel={'Bekreft ustending av frittstående brev'}
-                >
-                    {feilmelding && <Alert variant={'error'}>Utsending feilet.{feilmelding}</Alert>}
-                </ModalWrapper>
-            </Box>
-        );
-    } else if (utfall === 'OMGJØR_VEDTAK') {
-        return <OmgjørVedtak behandlingId={behandlingId} />;
-    } else {
-        return <div>{feilmelding || <SystemetLaster />}</div>;
+    if (feilmelding) {
+        return <Alert variant="error">{feilmelding}</Alert>;
     }
+
+    return (
+        <DataViewer response={{ vurdering }}>
+            {({ vurdering }) => {
+                if (vurdering && vurdering.vedtak) {
+                    if (vurdering.vedtak === VedtakValg.OPPRETTHOLD_VEDTAK) {
+                        return (
+                            <OpprettholdVedtak
+                                behandlingId={behandlingId}
+                                fagsystem={fagsystem}
+                                vurdering={vurdering}
+                            />
+                        );
+                    }
+
+                    if (vurdering.vedtak === VedtakValg.OMGJØR_VEDTAK) {
+                        return <OmgjørVedtak behandlingId={behandlingId} />;
+                    }
+                }
+                return <Alert variant="error">Noe gikk galt</Alert>;
+            }}
+        </DataViewer>
+    );
 };
